refactor: clean up unused files and update local store import
