--- static/scripts/demo/abis/index.ts
@@ -1,1 +1,0 @@
-export * from "./erc20Abi";
