+++ static/scripts/demo/local-store.ts
