+++ static/types/default-configuration.yml
